SLING-4716: delete distribution packages that are not imported correctly by importStream
git-svn-id: https://svn.apache.org/repos/asf/sling/trunk/contrib/extensions/distribution/core@1680244 13f79535-47bb-0310-9956-ffa450edef68

+++ src/main/java/org/apache/sling/distribution/packaging/impl/DistributionPackageUtils.java
@@ -46,8 +46,19 @@ public class DistributionPackageUtils {
                 log.error("package {} cannot be released from null queue", distributionPackage.getId());
             }
         } else {
-            distributionPackage.delete();
+            deleteSafely(distributionPackage);
             log.debug("package {} deleted", distributionPackage.getId());
+        }
+    }
+
+    public static void deleteSafely(DistributionPackage distributionPackage) {
+        if (distributionPackage == null) {
+            return;
+        }
+        try {
+            distributionPackage.delete();
+        } catch (Throwable t) {
+            log.error("error deleting package", t);
         }
     }
